+++ sample_app/sampleapp_common.c
@@ -134,7 +134,7 @@ static int app_write_ind(
    pnet_t                  *net,
    void                    *arg,
    uint32_t                arep,
-   uint16_t                api,
+   uint32_t                api,
    uint16_t                slot,
    uint16_t                subslot,
    uint16_t                idx,
@@ -204,7 +204,7 @@ static int app_read_ind(
    pnet_t                  *net,
    void                    *arg,
    uint32_t                arep,
-   uint16_t                api,
+   uint32_t                api,
    uint16_t                slot,
    uint16_t                subslot,
    uint16_t                idx,
@@ -368,7 +368,7 @@ static int app_reset_ind(
 static int app_exp_module_ind(
    pnet_t                  *net,
    void                    *arg,
-   uint16_t                api,
+   uint32_t                api,
    uint16_t                slot,
    uint32_t                module_ident)
 {
@@ -454,7 +454,7 @@ static int app_exp_module_ind(
 static int app_exp_submodule_ind(
    pnet_t                  *net,
    void                    *arg,
-   uint16_t                api,
+   uint32_t                api,
    uint16_t                slot,
    uint16_t                subslot,
    uint32_t                module_ident,
